# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "smirk-backend-core",
    "description": "Open, self-hostable backend for the Smirk non-custodial multi-chain wallet.",
    "contact": {
      "name": "John Winter Murphy",
      "email": "jw@such.software"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.3.0"
  },
  "paths": {
...
    "/wallet/grin/broadcast": {
      "post": {
        "tags": [
          "grin"
        ],
        "summary": "Broadcast a finalized, signed Grin transaction.",
        "operationId": "broadcast",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrinBroadcastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Transaction broadcast",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GrinBroadcastResponse"
                }
              }
            }
          },
          "400": {
            "description": "Grin disabled or malformed transaction"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/wallet/lws/balance": {
      "post": {
        "tags": [
          "xmr_wow"
        ],
        "summary": "Balance + scan state for an account.",
        "operationId": "balance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Account balance and scan state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LwsBalanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset, address, or view key"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/wallet/lws/history": {
      "post": {
        "tags": [
          "xmr_wow"
        ],
        "summary": "Transaction history (confirmed + mempool) for an account.",
        "operationId": "history",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ViewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Account transaction history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LwsHistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset, address, or view key"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/wallet/utxo/balance": {
      "post": {
        "tags": [
          "btc_ltc"
        ],
        "summary": "Confirmed/unconfirmed balance for a BTC/LTC address.",
        "operationId": "balance",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Address balance in satoshis",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BalanceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset or address"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/wallet/utxo/broadcast": {
      "post": {
        "tags": [
          "btc_ltc"
        ],
        "summary": "Broadcast a finalized, signed BTC/LTC transaction.",
        "operationId": "broadcast",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BroadcastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Broadcast accepted; returns the txid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BroadcastResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset or malformed tx hex"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/wallet/utxo/fee": {
      "post": {
        "tags": [
          "btc_ltc"
        ],
        "summary": "Fee-rate estimate (sat/vB) for confirmation within `blocks` blocks.",
        "operationId": "fee",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Fee estimate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeeResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/wallet/utxo/history": {
      "post": {
        "tags": [
          "btc_ltc"
        ],
        "summary": "Confirmed + mempool transaction history for a BTC/LTC address.",
        "operationId": "history",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Transaction history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset or address"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/wallet/utxo/tip": {
      "post": {
        "tags": [
          "btc_ltc"
        ],
        "summary": "Best-chain tip height (for client-side confirmation counting).",
        "operationId": "tip",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Chain tip height",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TipResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/wallet/utxo/utxos": {
      "post": {
        "tags": [
          "btc_ltc"
        ],
        "summary": "Unspent outputs for a BTC/LTC address.",
        "operationId": "utxos",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Unspent outputs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UtxosResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid/disabled asset or address"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "503": {
            "description": "Upstream node unavailable"
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AddressRequest": {
        "type": "object",
        "description": "An address query for a UTXO asset.",
        "required": [
          "asset",
          "address"
        ],
        "properties": {
          "address": {
            "type": "string",
            "description": "The address to query (validated by the Electrum client)."
          },
          "asset": {
            "type": "string",
            "description": "`btc` or `ltc`."
          }
        }
      },
...
      "AssetRequest": {
        "type": "object",
        "description": "An asset-only query.",
        "required": [
          "asset"
        ],
        "properties": {
          "asset": {
            "type": "string",
            "description": "`btc` or `ltc`."
          }
        }
      },
...
      "BalanceResponse": {
        "type": "object",
        "description": "Confirmed/unconfirmed balance in satoshis. The client computes any total\n(kept as separate integer fields — no lossy server-side sum).",
        "required": [
          "asset",
          "address",
          "confirmed",
          "unconfirmed"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "asset": {
            "type": "string"
          },
          "confirmed": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "unconfirmed": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "BroadcastRequest": {
        "type": "object",
        "description": "A broadcast request: a finalized, signed raw transaction.",
        "required": [
          "asset",
          "tx_hex"
        ],
        "properties": {
          "asset": {
            "type": "string",
            "description": "`btc` or `ltc`."
          },
          "tx_hex": {
            "type": "string",
            "description": "Hex-encoded signed transaction."
          }
        }
      },
      "BroadcastResponse": {
        "type": "object",
        "required": [
          "txid"
        ],
        "properties": {
          "txid": {
            "type": "string",
            "description": "The broadcast transaction id."
          }
        }
      },
...
      "FeeRequest": {
        "type": "object",
        "description": "A fee-estimate request.",
        "required": [
          "asset",
          "blocks"
        ],
        "properties": {
          "asset": {
            "type": "string",
            "description": "`btc` or `ltc`."
          },
          "blocks": {
            "type": "integer",
            "format": "int32",
            "description": "Target confirmation within this many blocks.",
            "minimum": 0
          }
        }
      },
      "FeeResponse": {
        "type": "object",
        "required": [
          "asset"
        ],
        "properties": {
          "asset": {
            "type": "string"
          },
          "sat_per_vb": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Estimated fee rate in sat/vB, or `null` if the server can't estimate."
          }
        }
      },
...
      "GrinBroadcastRequest": {
        "type": "object",
        "description": "Broadcast a finalized, signed Grin transaction (built + signed by the wallet).",
        "required": [
          "tx"
        ],
        "properties": {
          "tx": {
            "type": "object",
            "description": "The finalized transaction object (grin node `push_transaction` input)."
          }
        }
      },
      "GrinBroadcastResponse": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
...
      "HistoryEntry": {
        "type": "object",
        "description": "A transaction-history entry.",
        "required": [
          "txid",
          "height"
        ],
        "properties": {
          "fee": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Fee in satoshis (mempool entries only).",
            "minimum": 0
          },
          "height": {
            "type": "integer",
            "format": "int64",
            "description": "Block height (`0`/negative for unconfirmed)."
          },
          "txid": {
            "type": "string"
          }
        }
      },
      "HistoryResponse": {
        "type": "object",
        "required": [
          "asset",
          "address",
          "transactions"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "asset": {
            "type": "string"
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/HistoryEntry"
            }
          }
        }
      },
...
      "LwsBalanceResponse": {
        "type": "object",
        "description": "Balance + scan state. `total_received` is the balance (an LWS cannot confirm\nspends without the spend key; the wallet subtracts its own known spends).",
        "required": [
          "asset",
          "total_received",
          "unlocked",
          "locked",
          "scanned_height",
          "blockchain_height",
          "transaction_count"
        ],
        "properties": {
          "asset": {
            "type": "string"
          },
          "blockchain_height": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "locked": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "scanned_height": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total_received": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "transaction_count": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "unlocked": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "LwsHistoryResponse": {
        "type": "object",
        "required": [
          "asset",
          "transactions"
        ],
        "properties": {
          "asset": {
            "type": "string"
          },
          "transactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TxDto"
            }
          }
        }
      },
...
      "SpentOutputDto": {
        "type": "object",
        "description": "A candidate spent output (verify with the spend key before trusting).",
        "required": [
          "amount",
          "key_image",
          "tx_pub_key",
          "out_index",
          "mixin"
        ],
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "key_image": {
            "type": "string"
          },
          "mixin": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "out_index": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "tx_pub_key": {
            "type": "string"
          }
        }
      },
...
      "TipResponse": {
        "type": "object",
        "required": [
          "asset",
          "height"
        ],
        "properties": {
          "asset": {
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int64",
            "description": "Best-chain tip height."
          }
        }
      },
      "TxDto": {
        "type": "object",
        "description": "A transaction in the account's history.",
        "required": [
          "hash",
          "height",
          "timestamp",
          "total_received",
          "total_sent",
          "mempool",
          "unlock_time",
          "spent_outputs"
        ],
        "properties": {
          "hash": {
            "type": "string"
          },
          "height": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "mempool": {
            "type": "boolean"
          },
          "payment_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "spent_outputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SpentOutputDto"
            }
          },
          "timestamp": {
            "type": "string"
          },
          "total_received": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "total_sent": {
            "type": "integer",
            "format": "int64",
            "description": "\"Possible\" sent — candidate spends, not authoritative.",
            "minimum": 0
          },
          "unlock_time": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "Utxo": {
        "type": "object",
        "description": "A single unspent output.",
        "required": [
          "txid",
          "vout",
          "value",
          "height"
        ],
        "properties": {
          "height": {
            "type": "integer",
            "format": "int64",
            "description": "Block height; `0` if unconfirmed.",
            "minimum": 0
          },
          "txid": {
            "type": "string"
          },
          "value": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "vout": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "UtxosResponse": {
        "type": "object",
        "required": [
          "asset",
          "address",
          "utxos"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "asset": {
            "type": "string"
          },
          "utxos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Utxo"
            }
          }
        }
      },
      "ViewRequest": {
        "type": "object",
        "description": "Address + private view key for a per-account query.",
        "required": [
          "asset",
          "address",
          "view_key"
        ],
        "properties": {
          "address": {
            "type": "string"
          },
          "asset": {
            "type": "string",
            "description": "`xmr` or `wow`."
          },
          "view_key": {
            "type": "string",
            "description": "Private view key (64 hex). Forwarded to the LWS; never stored or logged."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Session access token issued by the /auth/* endpoints."
      }
    }
  },
...
}